Struct linux_object::process::LinuxProcess

pub struct LinuxProcess { /* fields omitted */ }

Linux specific process information.

Implementations

impl LinuxProcess

pub fn lookup_inode_at(
&self,
dirfd: FileDesc,
path: &str,
follow: bool
) -> LxResult<Arc<dyn INode>>

Lookup INode from the process.

If path is relative, then it is interpreted relative to the directory referred to by the file descriptor dirfd.

If the dirfd is the special value AT_FDCWD, then the directory is current working directory of the process.

If path is absolute, then dirfd is ignored.

If follow is true, then dereference path if it is a symbolic link.

pub fn lookup_inode(&self, path: &str) -> LxResult<Arc<dyn INode>>

Lookup INode from the process.

see lookup_inode_at

impl LinuxProcess

pub fn new(rootfs: Arc<dyn FileSystem>) -> Self

Create a new process.

pub fn get_futex(&self, uaddr: VirtAddr) -> Arc<Futex>

Get futex object.

pub fn add_file(&self, file: Arc<dyn FileLike>) -> LxResult<FileDesc>

Add a file to the file descriptor table.

pub fn add_file_at(&self, fd: FileDesc, file: Arc<dyn FileLike>)

Add a file to the file descriptor table at given fd.

pub fn get_file(&self, fd: FileDesc) -> LxResult<Arc<File>>

Get the File with given fd.

pub fn get_file_like(&self, fd: FileDesc) -> LxResult<Arc<dyn FileLike>>

Get the FileLike with given fd.

pub fn close_file(&self, fd: FileDesc) -> LxResult

Close file descriptor fd.

pub fn root_inode(&self) -> &Arc<dyn INode>

Get root INode of the process.

pub fn parent(&self) -> Option<Arc<Process>>

Get parent process.

pub fn current_working_directory(&self) -> String

Get current working directory.

pub fn change_directory(&self, path: &str)

Change working directory.

pub fn execute_path(&self) -> String

Get execute path.

pub fn set_execute_path(&self, path: &str)

Set execute path.