Struct linux_object::sync::Semaphore

pub struct Semaphore { /* fields omitted */ }

A counting, blocking, semaphore.

Implementations

impl Semaphore

pub fn new(count: isize) -> Semaphore

Creates a new semaphore with the initial count specified.

The count specified can be thought of as a number of resources, and a call to acquire or access will block until at least one resource is available. It is valid to initialize a semaphore with a negative count.

pub fn remove(&self)

Set the semaphore in removed statue

pub async fn acquire<'_>(&'_ self) -> Result<(), LxError>

Acquires a resource of this semaphore, blocking the current thread until it can do so.

This method will block until the internal count of the semaphore is at least 1.

pub fn release(&self)

Release a resource from this semaphore.

This will increment the number of resources in this semaphore by 1 and will notify any pending waiters in acquire or access if necessary.

pub async fn access<'_, '_>(&'_ self) -> Result<SemaphoreGuard<'_>, LxError>

Acquires a resource of this semaphore, returning an RAII guard to release the semaphore when dropped.

This function is semantically equivalent to an acquire followed by a release when the guard returned is dropped.

pub fn get(&self) -> isize

Get the current count

pub fn get_ncnt(&self) -> usize

Get the current eventbus callback length

pub fn get_pid(&self) -> usize

Get the current pid

pub fn set_pid(&self, pid: usize)

Set the current pid

pub fn set(&self, value: isize)

Set the current count