Trait linux_object::fs::vfs::INode

pub trait INode: Any + Sync + Send {

    fn read_at(&self, offset: usize, buf: &mut [u8]) -> Result<usize, FsError>;

    fn write_at(&self, offset: usize, buf: &[u8]) -> Result<usize, FsError>;

    fn poll(&self) -> Result<PollStatus, FsError>;

    fn as_any_ref(&self) -> &(dyn Any + 'static);

    fn async_poll(
        &'a self
    ) -> Pin<Box<dyn Future<Output = Result<PollStatus, FsError>> + Send + Sync + 'a, Global>> { ... }

    fn metadata(&self) -> Result<Metadata, FsError> { ... }

    fn set_metadata(&self, _metadata: &Metadata) -> Result<(), FsError> { ... }

    fn sync_all(&self) -> Result<(), FsError> { ... }

    fn sync_data(&self) -> Result<(), FsError> { ... }

    fn resize(&self, _len: usize) -> Result<(), FsError> { ... }

    fn create(
        &self, 
        name: &str, 
        type_: FileType, 
        mode: u32
    ) -> Result<Arc<dyn INode + 'static>, FsError> { ... }

    fn create2(
        &self, 
        name: &str, 
        type_: FileType, 
        mode: u32, 
        _data: usize
    ) -> Result<Arc<dyn INode + 'static>, FsError> { ... }

    fn link(
        &self, 
        _name: &str, 
        _other: &Arc<dyn INode + 'static>
    ) -> Result<(), FsError> { ... }

    fn unlink(&self, _name: &str) -> Result<(), FsError> { ... }

    fn move_(
        &self, 
        _old_name: &str, 
        _target: &Arc<dyn INode + 'static>, 
        _new_name: &str
    ) -> Result<(), FsError> { ... }

    fn find(&self, _name: &str) -> Result<Arc<dyn INode + 'static>, FsError> { ... }

    fn get_entry(&self, _id: usize) -> Result<String, FsError> { ... }

    fn get_entry_with_metadata(
        &self, 
        id: usize
    ) -> Result<(Metadata, String), FsError> { ... }

    fn io_control(&self, _cmd: u32, _data: usize) -> Result<usize, FsError> { ... }

    fn mmap(&self, _area: MMapArea) -> Result<(), FsError> { ... }

    fn fs(&self) -> Arc<dyn FileSystem + 'static> { ... }

}

Abstract file system object such as file or directory.

Required methods

fn read_at(&self, offset: usize, buf: &mut [u8]) -> Result<usize, FsError>

Read bytes at offset into buf, return the number of bytes read.

fn write_at(&self, offset: usize, buf: &[u8]) -> Result<usize, FsError>

Write bytes at offset from buf, return the number of bytes written.

fn poll(&self) -> Result<PollStatus, FsError>

Poll the events, return a bitmap of events.

fn as_any_ref(&self) -> &(dyn Any + 'static)

This is used to implement dynamics cast. Simply return self in the implement of the function.

Provided methods

fn async_poll(
&'a self
) -> Pin<Box<dyn Future<Output = Result<PollStatus, FsError>> + Send + Sync + 'a, Global>>

Poll the events, return a bitmap of events, async version.

fn metadata(&self) -> Result<Metadata, FsError>

Get metadata of the INode

fn set_metadata(&self, _metadata: &Metadata) -> Result<(), FsError>

Set metadata of the INode

fn sync_all(&self) -> Result<(), FsError>

Sync all data and metadata

fn sync_data(&self) -> Result<(), FsError>

Sync data (not include metadata)

fn resize(&self, _len: usize) -> Result<(), FsError>

Resize the file

fn create(
&self,
name: &str,
type_: FileType,
mode: u32
) -> Result<Arc<dyn INode + 'static>, FsError>

Create a new INode in the directory

fn create2(
&self,
name: &str,
type_: FileType,
mode: u32,
_data: usize
) -> Result<Arc<dyn INode + 'static>, FsError>

Create a new INode in the directory, with a data field for usages like device file.

fn link(
&self,
_name: &str,
_other: &Arc<dyn INode + 'static>
) -> Result<(), FsError>

Create a hard link name to other

fn unlink(&self, _name: &str) -> Result<(), FsError>

Delete a hard link name

fn move_(
&self,
_old_name: &str,
_target: &Arc<dyn INode + 'static>,
_new_name: &str
) -> Result<(), FsError>

Move INode self/old_name to target/new_name. If target equals self, do rename.

fn find(&self, _name: &str) -> Result<Arc<dyn INode + 'static>, FsError>

Find the INode name in the directory

fn get_entry(&self, _id: usize) -> Result<String, FsError>

Get the name of directory entry

fn get_entry_with_metadata(
&self,
id: usize
) -> Result<(Metadata, String), FsError>

Get the name of directory entry with metadata

fn io_control(&self, _cmd: u32, _data: usize) -> Result<usize, FsError>

Control device

fn mmap(&self, _area: MMapArea) -> Result<(), FsError>

Map files or devices into memory

fn fs(&self) -> Arc<dyn FileSystem + 'static>

Get the file system of the INode

Implementations

impl dyn INode + 'static

pub fn downcast_ref<T>(&self) -> Option<&T> where
T: INode,

Downcast the INode to specific struct

pub fn list(&self) -> Result<Vec<String, Global>, FsError>

Get all directory entries as a Vec

pub fn lookup(&self, path: &str) -> Result<Arc<dyn INode + 'static>, FsError>

Lookup path from current INode, and do not follow symlinks

pub fn lookup_follow(
&self,
path: &str,
follow_times: usize
) -> Result<Arc<dyn INode + 'static>, FsError>

Lookup path from current INode, and follow symlinks at most follow_times times