Struct linux_object::process::LinuxProcess

pub struct LinuxProcess { /* fields omitted */ }

Linux specific process information.

Implementations

impl LinuxProcess

pub fn lookup_inode_at(
&self,
dirfd: FileDesc,
path: &str,
follow: bool
) -> LxResult<Arc<dyn INode>>

Lookup INode from the process.

If path is relative, then it is interpreted relative to the directory referred to by the file descriptor dirfd.

If the dirfd is the special value AT_FDCWD, then the directory is current working directory of the process.

If path is absolute, then dirfd is ignored.

If follow is true, then dereference path if it is a symbolic link.

pub fn lookup_inode(&self, path: &str) -> LxResult<Arc<dyn INode>>

Lookup INode from the process.

see lookup_inode_at

impl LinuxProcess

pub fn new(rootfs: Arc<dyn FileSystem>) -> Self

Create a new process.

pub fn get_futex(&self, uaddr: VirtAddr) -> Arc<Futex>

Get futex object.

pub fn get_free_fd(&self) -> FileDesc

Get lowest free fd

pub fn get_free_fd_from(&self, start: usize) -> FileDesc

get the lowest available fd great than or equal to start.

pub fn add_file(&self, file: Arc<dyn FileLike>) -> LxResult<FileDesc>

Add a file to the file descriptor table.

pub fn add_file_at(
&self,
fd: FileDesc,
file: Arc<dyn FileLike>
) -> LxResult<FileDesc>

Add a file to the file descriptor table at given fd.

pub fn file_limit(&self, new_limit: Option<RLimit>) -> RLimit

get and set file limit number

pub fn get_file(&self, fd: FileDesc) -> LxResult<Arc<File>>

Get the File with given fd.

pub fn get_file_like(&self, fd: FileDesc) -> LxResult<Arc<dyn FileLike>>

Get the FileLike with given fd.

pub fn get_files(&self) -> LxResult<HashMap<FileDesc, Arc<dyn FileLike>>>

get all files

pub fn close_file(&self, fd: FileDesc) -> LxResult

Close file descriptor fd.

pub fn add_socket(
&self,
socket: Arc<Mutex<dyn Socket>>
) -> LxResult<SocketHandle>

Add a socket to the socket set at given SocketHandle.

pub fn get_socket(&self, fd: SocketHandle) -> LxResult<Arc<Mutex<dyn Socket>>>

Get the Socket with given fd.

pub fn close_socket(&self, fd: SocketHandle) -> LxResult

Close file descriptor fd.

pub fn root_inode(&self) -> &Arc<dyn INode>

Get root INode of the process.

pub fn parent(&self) -> Option<Arc<Process>>

Get parent process.

pub fn current_working_directory(&self) -> String

Get current working directory.

pub fn change_directory(&self, path: &str)

Change working directory.

pub fn execute_path(&self) -> String

Get execute path.

pub fn set_execute_path(&self, path: &str)

Set execute path.

pub fn signal_action(&self, signal: LinuxSignal) -> SignalAction

Get signal action.

pub fn set_signal_action(&self, signal: LinuxSignal, action: SignalAction)

Set signal action.

pub fn remove_cloexec_files(&self)

Close file that FD_CLOEXEC is set

pub fn semaphores_add(&self, array: Arc<SemArray>) -> usize

Insert a SemArray and return its ID

pub fn semaphores_get(&self, id: usize) -> Option<Arc<SemArray>>

Get an semaphore set by id

pub fn semaphores_add_undo(&self, id: usize, num: u16, op: i16)

Add an undo operation

pub fn semaphores_remove(&self, id: usize)

Remove an SemArray by ID

pub fn shm_get_id(&self, id: usize) -> Option<usize>

get ShmId from Virtual Addr

pub fn shm_get(&self, id: usize) -> Option<ShmIdentifier>

get the ShmIdentifier from shm_identifiers

pub fn shm_pop(&self, id: usize)

Delete the ShmIdentifier from shm_identifiers

pub fn shm_add(&self, shared_guard: Arc<Mutex<ShmGuard>>) -> usize

Insert the SharedGuard and return its ID

pub fn shm_set(&self, id: usize, shm_id: ShmIdentifier)

Set Virtual Addr for shared memory